requirements.txt
akshare==1.17.7
mootdx==0.11.7
pandas==2.3.0
tqdm==4.66.4
tushare==1.4.21
scipy==1.14.1
qq-botpy==1.2.1
python_dotenv==1.1.1
